Add configurations.autosetup
This will automatically import and call `setup()`.

It is useful to keep the boilerplate with using django-configurations to
a minimum, and also play nice with helpers like "isort" that would tear
the "import" and calling of "configurations.setup()" apart - which is
understandable.

The downside obviously is that importing this module _will_ have
side-effects, which is considered to be bad in general [1].

So this is open for discussion/feedback.
I could imagine doing the automatic call to `setup()` only when Django
is detected for example.

1: http://chrismorgan.info/blog/say-no-to-import-side-effects-in-python.html